Moving customise-table from uniprotkb to shared.
